restructure the folder
