+++ src/main/java/com/feeder/flashsale/dao/GoodsDao.java
@@ -1,0 +1,4 @@
+package com.feeder.flashsale.dao;
+
+public class GoodsDao {
+}
