+++ src/main/java/com/feeder/flashsale/entity/Goods.java
@@ -1,10 +1,12 @@
-package com.feeder.flashsale.beans;
+package com.feeder.flashsale.entity;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
+import org.springframework.stereotype.Repository;
 
 @Data
 @Accessors(chain = true)
+@Repository
 public class Goods {
 
     private String id;
