+++ src/main/java/com/feeder/flashsale/service/GoodsService.java
@@ -1,0 +1,7 @@
+package com.feeder.flashsale.service;
+
+import com.feeder.flashsale.entity.Goods;
+
+public interface GoodsService {
+    Goods getGoodsByID(long id);
+}
