+++ src/main/java/com/feeder/flashsale/service/impl/GoodsServiceImpl.java
@@ -1,0 +1,13 @@
+package com.feeder.flashsale.service.impl;
+
+import com.feeder.flashsale.entity.Goods;
+import com.feeder.flashsale.service.GoodsService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class GoodsServiceImpl implements GoodsService {
+    @Override
+    public Goods getGoodsByID(long id) {
+        return null;
+    }
+}
